requirements.txt
slack-bolt==1.14.3
requests==2.26.0
python-dotenv==0.19.0